+++ src/BunnyMate.php
@@ -11,6 +11,7 @@ use craft\events\DefineBehaviorsEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterUrlRulesEvent;
 use craft\services\Fs;
+use craft\services\Utilities;
 use craft\web\UrlManager;
 
 use vaersaagod\bunnymate\behaviors\VideoAssetBehavior;
@@ -19,6 +20,7 @@ use vaersaagod\bunnymate\models\Settings;
 use vaersaagod\bunnymate\services\Purge;
 use vaersaagod\bunnymate\services\Stream;
 use vaersaagod\bunnymate\services\Videos;
+use vaersaagod\bunnymate\utilities\VideoUpload;
 use vaersaagod\bunnymate\web\twig\BunnyMateExtension;
 
 use yii\base\Event;
@@ -81,6 +83,7 @@ class BunnyMate extends Plugin
         $this->_registerWebhookRoute();
         $this->_registerAssetUrlOverride();
         $this->_registerAssetCleanup();
+        $this->_registerUtilities();
 
         Craft::$app->onInit(static function () {
             Craft::$app->getView()->registerTwigExtension(new BunnyMateExtension());
@@ -271,6 +274,22 @@ class BunnyMate extends Plugin
                     return;
                 }
                 $this->getVideos()->deleteVideoForAsset($asset);
+            }
+        );
+    }
+
+    /**
+     * Registers BunnyMate's CP utilities.
+     *
+     * @return void
+     */
+    private function _registerUtilities(): void
+    {
+        Event::on(
+            Utilities::class,
+            Utilities::EVENT_REGISTER_UTILITIES,
+            static function (RegisterComponentTypesEvent $event) {
+                $event->types[] = VideoUpload::class;
             }
         );
     }